Разбивка JS кода на модули
